Add notes for each function, move helper linked list functions to file memoryFunc.c
