+++ memory.c
@@ -1,7 +1,7 @@
 #include "shell.h"
 
 /**
-* *_realloc -  reallocates a memory block using malloc and free
+* _realloc -  reallocates a memory block using malloc and free
 * @ptr: void pointer
 * @old_size: already allocated size
 * @new_size: new size to allocate
@@ -50,93 +50,12 @@ void *_realloc(void *ptr, unsigned int old_size, unsigned int new_size)
 	return (np);
 }
 
-int add_mem(void **p, save_mem **head)
-{
-	save_mem *newnode;
-
-	/*create node */
-	newnode = malloc(sizeof(save_mem));
-	if (newnode == NULL)
-		return (-1);
-	newnode->loc = *p;
-
-	if (!*head)
-	{
-		newnode->next = NULL;
-		*head = newnode;
-	}
-	else
-	{
-		/* prepend to LL, FIFO */
-		newnode->next = *head;
-		*head = newnode;
-	}
-
-	return (1);
-}
-
-int remove_child_mem(void **p, save_mem **head)
-{
-	save_mem *tmp, *hold;
-
-	tmp = *head;
-	while (tmp)
-	{
-		hold = tmp;
-		tmp = tmp->next;
-		if (hold->loc != NULL)
-			free(hold->loc);
-		free(hold);
-
-		if (*p == tmp->loc)
-			return (1);
-	}
-
-	return (0);
-}
-
-int remove_mem(void **p, save_mem **head)
-{
-	save_mem *tmp, *hold;
-
-	tmp = *head;
-	if (*p == NULL)
-	{
-		/* remove and free all nodes */
-		while (tmp)
-		{
-			hold = tmp;
-			tmp = tmp->next;
-			if (hold->loc != NULL)
-				free(hold->loc);
-			free(hold);
-		}
-		*head = NULL;
-		return (1);
-	}
-
-	/* remove only *p */
-	hold = tmp;
-	while (tmp)
-	{
-		if (tmp->loc == *p)
-		{
-			if (hold == tmp)
-				*head = tmp->next;
-			else
-				hold->next = tmp->next;
-
-			free(tmp->loc);
-			free(tmp);
-			return (1);
-		}
-		hold = tmp;
-		tmp = tmp->next;
-	}
-
-	return (0);
-}
-
+/**
+ * _ref_mem - Helper function that calls a memory add or remove function
+ * @p: pointer to save in node, or remove from linked list
+ * @action: string literal that determines which function to call
+ * Return: -1 on error, else return value returned from other function call
+ */
 int _ref_mem(void *p, char *action)
 {
 	static save_mem *head = NULL;
@@ -151,6 +70,11 @@ int _ref_mem(void *p, char *action)
 	return (-1);
 }
 
+/**
+ * _malloc - Malloc and save a pointer to malloc'd space for memory linked list
+ * @size: size to malloc
+ * Return: pointer to malloc or NULL on error
+ */
 void *_malloc(unsigned int size)
 {
 	void *p;
@@ -167,6 +91,10 @@ void *_malloc(unsigned int size)
 	return (p);
 }
 
+/**
+ * _free - Free based on pointer passed
+ * @ptr: pointer to malloc'd memory to free, or if NULL, free all memory nodes
+ */
 void _free(void *ptr)
 {
 	_ref_mem(ptr, "remove");
